Move tests to their own module files
