+++ eyelid-match-ops/src/primitives/poly/test.rs
@@ -1,0 +1,7 @@
+//! Tests for basic polynomial operations.
+
+#[cfg(any(test, feature = "benchmark"))]
+pub mod gen;
+
+#[cfg(test)]
+pub mod mul;
